% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{get}
\alias{get}
\alias{get,MolgenisSession-method}
\title{Retrieve data}
\usage{
get(session, entityTypeID, query = list())

\S4method{get}{MolgenisSession}(session, entityTypeID, query = list())
}
\arguments{
\item{session}{the current session}

\item{entityTypeID}{the entityTypeID to retrieve}

\item{query}{list with query parameters}
}
\value{
content of the response
}
\description{
Retrieve data
}
\section{Methods (by class)}{
\itemize{
\item \code{MolgenisSession}: Get using MolgenisSession
}}

