% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\docType{methods}
\name{login}
\alias{login}
\alias{login,MolgenisSession-method}
\title{Log in on a MOLGENIS server}
\usage{
login(session, username, password)

\S4method{login}{MolgenisSession}(session, username, password)
}
\arguments{
\item{session}{the current session}

\item{username}{the username to use to log in}

\item{password}{the password to use to log in}
}
\description{
Log in on a MOLGENIS server
}
\section{Methods (by class)}{
\itemize{
\item \code{MolgenisSession}: Log in using username and password
}}

