% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\docType{methods}
\name{logout}
\alias{logout}
\alias{logout,MolgenisSession-method}
\title{Logout}
\usage{
logout(session)

\S4method{logout}{MolgenisSession}(session)
}
\arguments{
\item{session}{the session to log out of}
}
\description{
Logout
}
\section{Methods (by class)}{
\itemize{
\item \code{MolgenisSession}: Logout session
}}

