% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.addAll}
\alias{molgenis.addAll}
\title{Creates new entities}
\usage{
molgenis.addAll(entity, rows)
}
\arguments{
\item{entity}{the entityname}

\item{rows}{dataFrame where each row is an entity}
}
\description{
Creates new entities
}
\examples{
firstName <- c("Piet", "Klaas")
lastName  <- c("Paulusma", "de Vries")
df <- data.frame(firstName, lastName)
molgenis.addAll("Person", df)

}
