% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.addList}
\alias{molgenis.addList}
\title{Creates a new entity}
\usage{
molgenis.addList(entity, attributeList)
}
\arguments{
\item{entity}{the entityname}

\item{attributeList}{list of attribute name/value pairs}
}
\value{
id of the created entity
}
\description{
Creates a new entity
}
