% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.delete}
\alias{molgenis.delete}
\title{Deletes an existing entity}
\usage{
molgenis.delete(entity, id)
}
\arguments{
\item{entity}{entityname}

\item{id}{the id of the entity to delete}
}
\description{
Deletes an existing entity
}
