% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.deleteList}
\alias{molgenis.deleteList}
\title{Deletes a list of entities in an entityType.}
\usage{
molgenis.deleteList(entity, rows)
}
\arguments{
\item{entity}{the entityType name}

\item{rows}{List with ids of the rows}
}
\description{
Deletes a list of entities in an entityType.
}
