% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.get}
\alias{molgenis.get}
\title{Query the data api}
\usage{
molgenis.get(entity, q = NULL, start = 0, num = 1000, sortColumn = NULL,
  sortOrder = NULL, attributes = NULL)
}
\arguments{
\item{entity}{the entityname (required)}

\item{q}{the query}

\item{start}{the index of the first result to return}

\item{num}{the number of results to return (default is 1000, max is 10000)}

\item{attributes}{the attributes to return if NULL (default) all attributes are returned}
}
\value{
Dataframe with the query result
}
\description{
The query must be in fiql/rsql format (see https://github.com/jirutka/rsql-parser)
}
\examples{
molgenis.get(entity = "Person", q = "firstName==Piet", start = 6, num = 10)
  
}
