% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.getAttributeMetaData}
\alias{molgenis.getAttributeMetaData}
\title{Gets attribute metadata}
\usage{
molgenis.getAttributeMetaData(entity, attribute)
}
\arguments{
\item{entity}{the entityname}

\item{attribute}{the attributename}
}
\value{
list of attributes
}
\description{
Gets attribute metadata
}
