% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.getEntityType}
\alias{molgenis.getEntityType}
\title{Gets entity metadata}
\usage{
molgenis.getEntityType(entity)
}
\arguments{
\item{entity}{the entityname}
}
\value{
list of entityTypes
}
\description{
Gets entity metadata
}
