% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.login}
\alias{molgenis.login}
\title{Login to the rest api and create a session}
\usage{
molgenis.login(api.url, username, password)
}
\arguments{
\item{api.url}{molgenis host}

\item{username}{your username}

\item{password}{your password}
}
\description{
Login on a MOLGENIS host to communicate with REST API
}
