% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.update}
\alias{molgenis.update}
\title{Updates an existing entity}
\usage{
molgenis.update(entity, id, ...)
}
\arguments{
\item{entity}{the entityname}

\item{id}{the id of the entity to update}

\item{var.arg}{list of attribute name/value pairs}
}
\description{
Updates an existing entity
}
\examples{
molgenis.update(entity = "Person", id = 5, firstName = "Piet", lastName = "Paulusma")

}
