% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisSession.R
\docType{class}
\name{MolgenisSession-class}
\alias{MolgenisSession-class}
\title{Class MolgenisSession.}
\description{
A Session on a MOLGENIS server.
}
\section{Slots}{

\describe{
\item{\code{handle}}{handle to access the server using httr}

\item{\code{token}}{authentication token, or "<NULL>" if not authenticated}
}}

