% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisSession.R
\name{Session}
\alias{Session}
\title{Create a MOLGENIS Session}
\usage{
Session(host)
}
\arguments{
\item{host}{the MOLGENIS hostname}
}
\value{
a [MolgenisSession] object.
}
\description{
Create a MOLGENIS Session
}
\examples{
Session("https://master.dev.molgenis.org/")

}
