% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRApi.R
\name{molgenis.login}
\alias{molgenis.login}
\title{Login to the rest api and create a session}
\usage{
molgenis.login(molgenis.host, username, password)
}
\arguments{
\item{molgenis.host}{MOLGENIS host url}

\item{username}{your username}

\item{password}{your password}
}
\value{
molgenis token for further usage
}
\description{
Login on a MOLGENIS host to communicate with REST API
}
